move labelService to workbench
